+++ app/src/main/res/layout/activity_non_member_mbti_choice.xml
@@ -5,6 +5,9 @@
 
     <data>
 
+        <variable
+            name="viewModel"
+            type="com.beside153.peopleinside.viewmodel.login.NonMemberMbtiChoiceViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -33,6 +36,7 @@
                     android:layout_height="match_parent"
                     android:background="@android:color/transparent"
                     android:contentDescription="@string/back_button"
+                    android:onClick="@{() -> viewModel.onBackButtonClick()}"
                     android:src="@drawable/ic_left_arrow" />
 
             </com.google.android.material.appbar.MaterialToolbar>
@@ -56,7 +60,8 @@
             android:layout_marginHorizontal="16dp"
             android:layout_marginBottom="8dp"
             android:background="@drawable/background_agree_button"
-            android:enabled="true"
+            android:enabled="@{viewModel.isCompleteButtonEnable}"
+            android:onClick="@{() -> viewModel.onCompleteButtonClick()}"
             android:text="@string/complete_choose"
             android:textSize="18sp"
             app:backgroundTint="@null"
